+++ src/main/java/org/tnh/controllers/LoginController.java
@@ -57,11 +57,11 @@ public class LoginController
             loginMessage.setText(e.getMessage());
         }
     }
-//de modificat
+
     public void handleShowRecipes(ActionEvent event) throws Exception {
-        root = FXMLLoader.load(Objects.requireNonNull(getClass().getClassLoader().getResource("recipe.fxml")));
+        root = FXMLLoader.load(Objects.requireNonNull(getClass().getClassLoader().getResource("list_recipes.fxml")));
         stage = (Stage)((Node)event.getSource()).getScene().getWindow();
-        stage.setTitle("Recipes");
+        stage.setTitle("List of recipes");
         stage.setScene(new Scene(root, 900, 550));
         stage.show();
     }
