+++ src/main/resources/create_recipe.fxml
@@ -37,7 +37,7 @@
     <GridPane id="thirdGrid" GridPane.rowIndex="1" GridPane.columnIndex="0" vgap="10" hgap="10">
         <Text fx:id="recipeMessage" GridPane.columnIndex="1" GridPane.rowIndex="5"/>
         <HBox spacing="10" GridPane.columnIndex="1" GridPane.rowIndex="3">
-            <Button fx:id="register_button" onAction="#handleRegisterAction" text="Register"/>
+            <Button fx:id="register_button" onAction="#handleRegisterAction" text="Create"/>
             <Button fx:id="back_button" onAction="#handleBackFirstPage" text="Back"/>
         </HBox>
     </GridPane>
