+++ src/main/resources/head.fxml
@@ -15,8 +15,8 @@
         <padding>
             <Insets top="25.0" bottom="25.0" left="25.0" right="25.0"/>
         </padding>
-        <Button fx:id ="create_button" text="Create recipe" GridPane.rowIndex="0" GridPane.columnIndex="0"/>
-        <Button fx:id ="recipes_button" text="Show recipes" GridPane.rowIndex="1" GridPane.columnIndex="0"/>
+        <Button fx:id ="create_button" onAction="#handleCreateRecipeAction" text="Create recipe" GridPane.rowIndex="0" GridPane.columnIndex="0"/>
+        <Button fx:id ="recipes_button" onAction="#handleShowRecipes" text="Show recipes" GridPane.rowIndex="1" GridPane.columnIndex="0"/>
         <Button fx:id ="logout_button" onAction="#handleLogout" text="Logout" alignment="TOP_CENTER" GridPane.rowIndex="35" GridPane.columnIndex="0"/>
     </GridPane>
     <GridPane id="searchGrid" GridPane.rowIndex="0" GridPane.columnIndex="1" vgap="10" hgap="10">
