+++ src/main/resources/list_recipes.fxml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.geometry.Insets?>
+
+<GridPane fx:id="mainGrid"
+          fx:controller="org.tnh.controllers.ListRecipesController"
+          xmlns:fx="http://javafx.com/fxml"
+          vgap="10" hgap="220">
+
+    <GridPane id="buttonsGrid" GridPane.rowIndex="0" GridPane.columnIndex="0" vgap="10" hgap="10">
+        <padding>
+            <Insets top="25.0" bottom="25.0" left="25.0" right="25.0"/>
+        </padding>
+        <Button fx:id ="logout_button" onAction="#handleBackFirstPage" text="Logout" alignment="TOP_LEFT" GridPane.rowIndex="0" GridPane.columnIndex="0"/>
+    </GridPane>
+
+</GridPane>
